fix: ignore paths that begin with underscore, since they cannot be stored in PouchDB anyway

+++ src/scanner.ts
@@ -39,6 +39,11 @@ export default function (logger: Logger, db: MediaDatabase, config: Record<strin
 		.pipe(
 			concatMap(async ([mediaPath, mediaStat]) => {
 				const mediaId = getId(config.paths.media, mediaPath)
+
+				// Filter out files that are not to be scanned
+				if (!filter(mediaId, mediaPath)) return
+
+				console.log('mediaId', mediaId)
 				try {
 					if (!mediaStat) {
 						await db.remove(await db.get(mediaId))
@@ -147,5 +152,17 @@ export default function (logger: Logger, db: MediaDatabase, config: Record<strin
 		await db.put(doc, { force: true })
 
 		mediaLogger.info('Scanned')
+	}
+	/**
+	 * Filter out files that are not to be scanned
+	 * @returns true if the media is to be scanned, false otherwise
+	 */
+	function filter(mediaId: string, _mediaPath: string): boolean {
+		if (mediaId.startsWith('_')) {
+			// PouchDB cannot store these ("Only reserved document ids may start with underscore.")
+			return false
+		}
+
+		return true
 	}
 }
